Fix the error in pr #69
Now the buy button successfully redirect to claimEvent.
The Toolbar was updated to reflect the student buttons and the
respective paths were corrected.
Fix the pathing issue where when clicking on the home button, would send
to the organizer side.
Log out button seems to work as intended and the bug failed to be
reproduced.